backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "httpx[http2]>=0.28.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.41.0",
    "redis>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
